+++ src/dinpro/domain/linear_referencing/__init__.py
@@ -1,3 +1,4 @@
+from dinpro.domain.linear_referencing.dynamic_segmentation import DynamicSegmentation
 from dinpro.domain.linear_referencing.linear_referencing import LinearReferencing
 from dinpro.domain.linear_referencing.measure_system import (
     CalibrationPoint,
@@ -5,6 +6,7 @@ from dinpro.domain.linear_referencing.measure_system import (
     MeasureSystem,
 )
 from dinpro.domain.linear_referencing.pk import PK
+from dinpro.domain.linear_referencing.segment import Segment
 from dinpro.domain.linear_referencing.station import Station
 from dinpro.domain.linear_referencing.station_formatter import (
     ClassicFormatter,
@@ -18,10 +20,12 @@ from dinpro.domain.linear_referencing.station_parser import StationParser
 __all__ = [
     "PK",
     "Station",
+    "Segment",
     "LinearReferencing",
     "MeasureSystem",
     "CalibrationPoint",
     "MeasureDiscontinuity",
+    "DynamicSegmentation",
     "StationParser",
     "StationFormatter",
     "ClassicFormatter",